requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Kafka and streaming
kafka-python>=2.0.0

# Redis
redis>=4.0.0

# Web framework
Flask>=2.0.0
Flask-CORS>=3.0.0

# Data processing
python-dateutil>=2.8.0

# Utilities
python-dotenv>=0.19.0
requests>=2.25.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
flake8>=3.9.0

# Jupyter notebooks
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Monitoring and logging
structlog>=21.0.0
